move SearchLineEdit into custom_line_edit file
