+++ cockatrice/src/deck/custom_line_edit.cpp
@@ -6,6 +6,7 @@
 #include <QKeyEvent>
 #include <QLineEdit>
 #include <QObject>
+#include <QTreeView>
 #include <QWidget>
 
 LineEditUnfocusable::LineEditUnfocusable(QWidget *parent) : QLineEdit(parent)
@@ -68,4 +69,20 @@ bool LineEditUnfocusable::eventFilter(QObject *watched, QEvent *event)
     }
 
     return QLineEdit::eventFilter(watched, event);
+}
+
+void SearchLineEdit::keyPressEvent(QKeyEvent *event)
+{
+    // List of key events that must be handled by the card list instead of the search box
+    static const QVector<Qt::Key> forwardToTreeView = {Qt::Key_Up, Qt::Key_Down, Qt::Key_PageDown, Qt::Key_PageUp};
+    // forward only if the search text is empty
+    static const QVector<Qt::Key> forwardWhenEmpty = {Qt::Key_Home, Qt::Key_End};
+    Qt::Key key = static_cast<Qt::Key>(event->key());
+    if (treeView) {
+        if (forwardToTreeView.contains(key))
+            QCoreApplication::sendEvent(treeView, event);
+        if (text().isEmpty() && forwardWhenEmpty.contains(key))
+            QCoreApplication::sendEvent(treeView, event);
+    }
+    LineEditUnfocusable::keyPressEvent(event);
 }
